administration/..., base/...) : add a team setting card to user setting

+++ app/base/routes.py
@@ -4,6 +4,7 @@ Copyright (c) 2019 - present AppSeed.us
 Copyright (c) 2021 - present Orange Cyberdefense
 """
 
+import json
 import ssl
 from datetime import datetime
 from os import path
@@ -11,9 +12,10 @@ from os import path
 from app import db, ldap_manager, login_manager
 from app.administration.models import LdapConfiguration
 from app.administration.util import validate_user_form
+from app.administration.forms import UserForm
 from app.base import blueprint, util
 from app.base.forms import LoginForm, CreateUserForm, CreateTeamForm
-from app.base.models import User, Team
+from app.base.models import User, Team, team_members_association
 from app.base.util import (init_db, last_12_months_analysis_count,
                            ldap_config_dict, remove_dir_content, verify_pass)
 from app.projects.util import get_user_projects_ids
@@ -26,6 +28,7 @@ from flask import (current_app, redirect, render_template, request, session,
 from flask_login import current_user, login_required, login_user, logout_user
 from is_safe_url import is_safe_url
 from ldap3 import Tls
+from sqlalchemy.orm import aliased
 
 
 @blueprint.route("/")
@@ -217,9 +220,15 @@ def logout():
 @login_required
 def team_remove(team_id):
     team = Team.query.filter_by(id=team_id).first_or_404()
-    db.session.delete(team)
-    db.session.commit()
-    return redirect(url_for("base_blueprint.teams_setting"))
+    admin = util.is_admin(current_user.role)
+    if current_user.username == team.creator:
+        db.session.delete(team)
+        db.session.commit()
+        if admin:
+            return redirect(url_for("base_blueprint.teams_setting"))
+        else:
+            return redirect(url_for("base_blueprint.user_edit"))
+    return render_template("403.html"), 403
 
 @blueprint.route("teams_setting/edit/<team_id>", methods=["GET", "POST"])
 @login_required
@@ -227,7 +236,9 @@ def team_edit(team_id):
     edit_team = Team.query.filter_by(id=team_id).first_or_404()
     teamForm = CreateTeamForm()
     admin = util.is_admin(current_user.role)
-    if admin:
+    # check if acces is authorized
+    if current_user.username == edit_team.creator or admin:
+        # get sif is there is filter for members or projects
         member_name = request.form.get('member_name', '')
         project_name = request.form.get('project_name', '')
         if teamForm.members.data:
@@ -243,6 +254,7 @@ def team_edit(team_id):
             team=edit_team,
             member_name=member_name,
             project_name=project_name,
+            admin=admin,
             )
         if request.form.get('sort_project_button'):
             return render_template(
@@ -253,6 +265,7 @@ def team_edit(team_id):
             team=edit_team,
             member_name=member_name,
             project_name=project_name,
+            admin=admin,
             )
         if len(request.form) > 0:
             if teamForm.validate_on_submit():
@@ -264,9 +277,11 @@ def team_edit(team_id):
                 db.session.commit()
                 current_app.logger.info("Team updated (team.id=%i)", team_id)
                 flash("Team successfully updated", "success")
-                return redirect(url_for("base_blueprint.teams_setting"))
+                if admin:
+                    return redirect(url_for("base_blueprint.teams_setting"))
+                else:
+                    return redirect(url_for("base_blueprint.user_edit"))
             else:
-                print(teamForm.errors)
                 return render_template(
                 "team_edit.html",
                 edit=True,
@@ -275,6 +290,7 @@ def team_edit(team_id):
                 team=edit_team,
                 member_name=member_name,
                 project_name=project_name,
+                admin=admin,
             )
         else:
             teamForm.name.data = edit_team.name
@@ -286,6 +302,7 @@ def team_edit(team_id):
                 team=edit_team,
                 member_name=member_name,
                 project_name=project_name,
+                admin=admin,
             )
     else:
         return render_template("403.html"), 403
@@ -293,44 +310,8 @@ def team_edit(team_id):
 @blueprint.route("teams_setting/add", methods=["POST", "GET"])
 @login_required
 def team_add():
-    teamForm = CreateTeamForm()
     admin = util.is_admin(current_user.role)
-    if admin:
-        if request.method == "POST" and teamForm.validate_on_submit():
-            new_team = Team(
-                name=teamForm.name.data,
-                creator=current_user.username,
-                user_id=current_user.id
-            )
-            member_ids = teamForm.members.data
-            members = User.query.filter(User.id.in_(member_ids)).all()
-
-            new_team.members = members
-            project_ids = teamForm.projects.data
-            projects = Project.query.filter(Project.id.in_(project_ids)).all()
-
-            new_team.projects = projects
-            db.session.add(new_team)
-            db.session.commit()
-            return redirect(url_for("base_blueprint.teams_setting"))
-        return render_template("team_edit.html", user=current_user, form=teamForm, edit=False, team=None, member_name="", project_name="")
-    else:
-        return render_template("403.html"), 403
-
-@blueprint.route("/teams_setting", methods=["GET", "POST"])
-@login_required
-def teams_setting():
-    if not Team.query.filter(Team.name == 'Global').first():
-        global_team = Team(
-                name="Global",
-                creator=None,
-                user_id=None
-            )
-        global_team.members = db.session.query(User).all()
-        db.session.add(global_team)
-        db.session.commit()
     teamForm = CreateTeamForm()
-    teamTable = Team.query.all()
     if request.method == "POST" and teamForm.validate_on_submit():
         new_team = Team(
             name=teamForm.name.data,
@@ -347,8 +328,49 @@ def teams_setting():
         new_team.projects = projects
         db.session.add(new_team)
         db.session.commit()
-        return redirect(url_for("base_blueprint.teams_setting", edit=False))
-    return render_template("teams_setting.html", user=current_user, segment="teams", form=teamForm, teamTable=teamTable, edit=False)
+        if admin:
+            return redirect(url_for("base_blueprint.teams_setting"))
+        else:
+            return redirect(url_for("base_blueprint.user_edit"))
+    return render_template("team_edit.html", user=current_user, form=teamForm, edit=False, team=None, member_name="", project_name="")
+
+@blueprint.route("/teams_setting", methods=["GET", "POST"])
+@login_required
+def teams_setting():
+    admin = util.is_admin(current_user.role)
+    if admin:
+        if not Team.query.filter(Team.name == 'Global').first():
+            global_team = Team(
+                    name="Global",
+                    creator=None,
+                    user_id=None
+                )
+            global_team.members = db.session.query(User).all()
+            db.session.add(global_team)
+            db.session.commit()
+        teamForm = CreateTeamForm()
+        teamTable = Team.query.all()
+        if request.method == "POST" and teamForm.validate_on_submit():
+            new_team = Team(
+                name=teamForm.name.data,
+                creator=current_user.username,
+                user_id=current_user.id
+            )
+            member_ids = teamForm.members.data
+            members = User.query.filter(User.id.in_(member_ids)).all()
+
+            new_team.members = members
+            project_ids = teamForm.projects.data
+            projects = Project.query.filter(Project.id.in_(project_ids)).all()
+
+            new_team.projects = projects
+            db.session.add(new_team)
+            db.session.commit()
+            return redirect(url_for("base_blueprint.teams_setting", edit=False))
+        return render_template("teams_setting.html", user=current_user, segment="teams", form=teamForm, teamTable=teamTable, edit=False)
+    else :
+        return render_template("403.html"), 403
+
 
 
 
@@ -391,6 +413,80 @@ def index():
         user=current_user,
         segment="dashboard",
     )
+
+@blueprint.route("/user_edit", methods=["GET", "POST"])
+@login_required
+def user_edit() :
+    admin = util.is_admin(current_user.role)
+    edit_user = User.query.filter_by(id=current_user.id).first()
+    user_teams = (
+        Team.query
+        .join(team_members_association)
+        .join(User, team_members_association.c.user_id == User.id)
+        .filter(User.id == current_user.id)
+        .all()
+    )
+    # POST / Form submitted
+    if "save-user" in request.form:
+        user_form = UserForm()
+        # Form is valid
+        if user_form.validate_on_submit():
+            # Perform additional custom validation
+            err = validate_user_form(
+                form=user_form,
+                skip_username=(edit_user.username == user_form.username.data),
+                skip_email=(edit_user.email == user_form.email.data),
+                skip_password=(user_form.password.data == ""),
+            )
+            if err is not None:
+                flash(err, "error")
+                return render_template(
+                    "user_edit.html",
+                    edit=True,
+                    form=UserForm(obj=edit_user),
+                    user=current_user,
+                    segment="users",
+                    teams=user_teams,
+                    admin=admin
+                )
+            # Change the password if needed only
+            if user_form.password.data == "":
+                edit_password = edit_user.password
+            else:
+                edit_password = util.hash_pass(user_form.password.data)
+            # User can be updated
+            user_form.populate_obj(edit_user)
+            edit_user.password = edit_password
+            db.session.commit()
+            current_app.logger.info("User updated (user.id=%i)", user_form.id.data)
+            flash("User successfully updated", "success")
+            return redirect(url_for("base_blueprint.index"))
+        # Form is invalid, form.error is populated
+        else:
+            current_app.logger.warning(
+                "User edit form invalid entries: %s", json.dumps(user_form.errors)
+            )
+            flash(str(user_form.errors), "error")
+            return render_template(
+                "user_edit.html",
+                edit=True,
+                form=UserForm(obj=edit_user),
+                user=current_user,
+                segment="users",
+                teams=user_teams,
+                admin=admin,
+            )
+    # GET / Display form
+    else:
+        return render_template(
+            "user_edit.html",
+            edit=True,
+            form=UserForm(obj=edit_user),
+            user=current_user,
+            segment="users",
+            teams=user_teams,
+            admin=admin,
+        )
 
 
 # Errors
